utilities: Move `vm_console_run_commands` to the console module
